Manually cherry-pick relevant changes from ECO-1700 that should be in ECO-1718
