+++ src/typescript/sdk/src/emojicoin_dot_fun/utils.ts
@@ -10,7 +10,7 @@ import { EMOJICOIN_DOT_FUN_MODULE_NAME, MODULE_ADDRESS } from "../const";
 import { type Events, converter, toGenericEvent } from "./events";
 import { type ContractTypes } from "../types";
 import { TYPE_TAGS } from "../utils/type-tags";
-import { createNamedObjectAddress } from "../utils/misc";
+import { createNamedObjectAddress } from "../utils/aptos-utils";
 
 /**
  * Derives the object address from the given emoji hex codes vector<u8> seed and
