+++ src/typescript/sdk/src/markets/utils.ts
@@ -8,11 +8,26 @@ import {
   type HexInput,
   type InputGenerateTransactionOptions,
   parseTypeTag,
+  type TypeTag,
 } from "@aptos-labs/ts-sdk";
 import { EmojicoinDotFun, deriveEmojicoinPublisherAddress } from "../emojicoin_dot_fun";
 import { type EmojicoinInfo } from "../types/contract";
-import { toConfig } from "../utils/misc";
+import { toConfig } from "../utils/aptos-utils";
 import { COIN_FACTORY_MODULE_NAME, DEFAULT_REGISTER_MARKET_GAS_OPTIONS } from "../const";
+
+export function toCoinTypes(inputAddress: AccountAddressInput): {
+  emojicoin: TypeTag;
+  emojicoinLP: TypeTag;
+} {
+  const marketAddress = AccountAddress.from(inputAddress);
+
+  return {
+    emojicoin: parseTypeTag(`${marketAddress.toString()}::${COIN_FACTORY_MODULE_NAME}::Emojicoin`),
+    emojicoinLP: parseTypeTag(
+      `${marketAddress.toString()}::${COIN_FACTORY_MODULE_NAME}::EmojicoinLP`
+    ),
+  };
+}
 
 /**
  * Get the derived market address and TypeTags for the given registry address and symbol bytes.
@@ -32,12 +47,12 @@ export function getEmojicoinMarketAddressAndTypeTags(args: {
     emojis: [symbolBytes.toStringWithoutPrefix()],
   });
 
+  const { emojicoin, emojicoinLP } = toCoinTypes(marketAddress);
+
   return {
     marketAddress,
-    emojicoin: parseTypeTag(`${marketAddress.toString()}::${COIN_FACTORY_MODULE_NAME}::Emojicoin`),
-    emojicoinLP: parseTypeTag(
-      `${marketAddress.toString()}::${COIN_FACTORY_MODULE_NAME}::EmojicoinLP`
-    ),
+    emojicoin,
+    emojicoinLP,
   };
 }
 
@@ -71,22 +86,4 @@ export const registerMarketAndGetEmojicoinInfo = async (args: {
   });
 
   return { marketAddress, emojicoin, emojicoinLP };
-};
-
-export const divideWithPrecision = (args: {
-  a: bigint | number;
-  b: bigint | number;
-  decimals?: number;
-}): number => {
-  const decimals = args.decimals ?? 3;
-  const a = BigInt(args.a);
-  const b = BigInt(args.b);
-  const f = BigInt(10 ** decimals);
-  return Number((a * f) / b) / Number(f);
-};
-
-export const truncateAddress = (input: AccountAddressInput): string => {
-  const t = AccountAddress.from(input);
-  const s = t.toString();
-  return `${s.substring(0, 6)}...${s.substring(s.length - 4, s.length)}`;
 };
