+++ src/typescript/sdk/tsconfig.json
@@ -33,7 +33,8 @@
     ]
   },
   "exclude": [
-    "node_modules"
+    "node_modules",
+    "./src/markets/spam.ts"
   ],
   "include": [
     "src"
